Video2Text/requirements.txt
requests==2.31.0
lxml==4.9.3
moviepy==1.0.3
openai-whisper==20231117
torch==2.1.2
tqdm==4.66.1
opencc-python-reimplemented==0.1.7
numpy==1.24.3
ffmpeg-python==0.2.0
python-dotenv==1.0.1
tenacity==8.2.3 